KPI/kpi/requirements.txt
absl-py==1.0.0
actionlib==1.13.2
addict==2.4.0
-e git+https://github.com/AMZ-Driverless/autonomous_2022.git@b7133b550269183b396b02ea87dd19fd803cec07#egg=amz_kpi&subdirectory=tools/kpi
amz-tool==2021.11.6
angles==1.9.13
anyio==3.5.0
apturl==0.5.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.3
asttokens==2.0.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.0.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.24
catkin-pkg-modules==0.4.24
cbor==1.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cpplint==1.6.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
deprecation==2.1.0
descartes==1.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.2
empy==3.3.2
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastimport==0.9.8
ffmpeg-python==0.2.0
filelock==3.6.0
fonttools==4.29.1
future==0.18.2
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gnupg==2.3.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
gpg===1.13.1-unknown
grpcio==1.44.0
httplib2==0.14.0
hyperlink==19.0.0
identify==2.4.11
idna==3.3
image-geometry==1.16.0
importlib-metadata==4.11.2
importlib-resources==5.4.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.11.1
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
jupyterlab-widgets==1.1.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
message-filters==1.15.14
mistune==0.8.4
mizani==0.7.3
monotonic==1.5
mpi4py==3.0.3
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
nodeenv==1.6.0
nose==1.3.7
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.2
oauthlib==3.1.0
olefile==0.46
open3d==0.13.0
opencv-python==4.5.5.62
packaging==21.3
palettable==3.3.0
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
patsy==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
plotnine==0.8.0
pre-commit==2.17.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pyerf==1.0.1
PyGithub==1.43.7
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.12.2
pymacaroons==0.13.0
pymap3d==2.7.2
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pypng==0.0.20
pyproj==2.6.0
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2021.3
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.14
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospkg-modules==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.2
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.8
rviz==1.14.14
scikit-learn==1.0.2
scikits.bootstrap==1.1.0
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
testpath==0.6.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.10.0
topic-tools==1.15.14
torch==1.9.1
torchgeometry==0.1.2
torchvision==0.10.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.1.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.7
vcstools==0.1.42
virtualenv==20.13.2
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.13.3
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.13
xkit==0.0.0
yapf==0.32.0
yarg==0.1.9
zipp==3.7.0
zope.interface==4.7.1
